fix: Rename 使用非指定的模式可能产生的报错
