+++ .github/workflows/pull-request.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Generate pull request body
         id: pr
         run: |
-          echo "::set-output name=body::$(git log --format='* %s' $(git describe --tags --abbrev=0)..HEAD)"
+          echo "docs/changelog/${{ github.ref }}.md"
       - name: Create Pull Request
         uses: peter-evans/create-pull-request@v3
         with:
