+++ docs/changelog/3.0.md
@@ -22,8 +22,8 @@
 
 ### 通知模块
 
-- 新增 `Bark` 通知模块，可以使用 `Bark` 推送通知到手机。
-- 全新的通知方式，现在可以推送海报，番剧名称，更新集数。
+- 新增 `Bark` 通知模块.
+- 全新的通知方式，现在可以给 Telegram 推送海报，番剧名称，更新集数。
 
 ### 数据迁移
 
@@ -42,4 +42,5 @@
   - 加速启动关闭时间
 - 重构解析器模块
 - 重构重命名模块
+  - 暂时移除 `normal` 模式
 - 增加 `ghcr.io` 镜像
