+++ src/module/rss/__init__.py
@@ -1,4 +1,4 @@
-from .rss_analyser import RSSAnalyser
+from .analyser import RSSAnalyser
 from .add_rules import add_rules
